+++ tests/Feature/Factories/TrackActivityFactoryTest.php
@@ -2,6 +2,8 @@
 
 namespace Sfneal\Tracking\Tests\Feature\Factories;
 
+use Sfneal\Address\Models\Address;
+use Sfneal\Testing\Models\People;
 use Sfneal\Testing\Utils\Interfaces\Factory\FillablesTest;
 use Sfneal\Tracking\Models\TrackActivity;
 
@@ -19,9 +21,11 @@ class TrackActivityFactoryTest extends FactoriesTestCase implements FillablesTes
         $this->assertIsInt($this->model->user_id);
         $this->assertIsString($this->model->route);
         $this->assertIsString($this->model->description);
-        $this->assertIsString($this->model->model_table);
-        $this->assertIsInt($this->model->model_key);
         $this->assertIsArray($this->model->model_changes);
         $this->assertIsString($this->model->request_token);
+        $this->assertIsInt($this->model->trackable_id);
+        $this->assertIsString($this->model->trackable_type);
+        $this->assertContains($this->model->trackable_type, [People::class, Address::class]);
+        $this->assertInstanceOf($this->model->trackable_type, $this->model->trackable);
     }
 }